+++ Assets/Mirage/Runtime/ServerObjectManager.cs
@@ -103,6 +103,8 @@ namespace Mirage
 
         void OnFinishedSceneChange(string scenePath, SceneOperation sceneOperation)
         {
+            Server.World.RemoveDestroyedObjects();
+
             SpawnOrActivate();
         }