+++ Assets/Mirage/Runtime/SocketLayer/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 using System.Runtime.CompilerServices;
 
-[assembly: AssemblyVersion("105.1.0")]
+[assembly: AssemblyVersion("107.1.1")]
 
 [assembly: InternalsVisibleTo("Mirage.SocketLayer.Tests")]